+++ Civi/ActionProvider/Action/Communication/CreatePdf.php
@@ -54,7 +54,7 @@ class CreatePdf extends AbstractAction {
       $pdfFormat = $parameters->getParameter('page_format_id');
     }
 
-    $processedMessage = Tokens::replaceTokens($contactId, $message, $contact);
+    $processedMessage = Tokens::replaceTokens($contactId, $message, $contact, 'text/html');
     if ($processedMessage === false) {
       return;
     }