+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 1.91 (not yet released)
 ------------
 
+* Convert tokens with the token processor (only if civicrm is newer then version 5.42).
+
 Version 1.90
 ------------